Changed Floats to Doubles

+++ src/org/usfirst/frc/team1351/robot/Drive/DriveAuton.java
@@ -6,11 +6,11 @@ import edu.wpi.first.wpilibj.PIDController;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
 public class DriveAuton {
-	private static final float[] TICKSPERINCH = {487.5f, 487.5f}; //Ticks per inch values (Low - 0, High - 1) TODO Fix Values
-	private static final float TALONMOVETHRESHOLD = 2.5f;
-	private static final float TALONTURNTHRESHOLD = 0.5f;
-	private static final float MOVETHRESHOLD = 10f;
-	private static final float TURNTHRESHOLD = 3f;
+	private static final double[] TICKSPERINCH = {487.5f, 487.5f}; //Ticks per inch values (Low - 0, High - 1) TODO Fix Values
+	private static final double TALONMOVETHRESHOLD = 2.5f;
+	private static final double TALONTURNTHRESHOLD = 0.5f;
+	private static final double MOVETHRESHOLD = 10f;
+	private static final double TURNTHRESHOLD = 3f;
 
 	private static double proportionalConstant;
 	private static double integralConstant;
